pyproject.toml
[project]
name = "Bible-QnA-API"
version = "0.1.0"
description = "API For Bible Q&A website"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=45.0.3",
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
]
